+++ test/TypesSpec.hs
@@ -36,8 +36,8 @@ tests =
         , testCase "omits services key" $
             assertBool "unexpected services key" $
                 not ("\"services\"" `BS.isInfixOf` BSL.toStrict (encode emptyDoc))
-        , testCase "currentIRVersion is 8" $
-            currentIRVersion @?= IRVersion 8
+        , testCase "currentIRVersion is 9" $
+            currentIRVersion @?= IRVersion 9
         , testGroup
             "properties"
             [ testProperty "profileNameText . mkProfileName roundtrips non-empty text" $